feat: move components table to modules and fix some type error on city form
