+++ src/pages/tech-admin/clients/[clientId]/branches/query.graphql
@@ -1,0 +1,11 @@
+query branches {
+    branches {
+        id
+        name
+        description
+        client {
+            id
+            name
+        }
+    }
+}
